Added an assets folder, to be used for images and sounds. Added function for graphics to use images. Tested movement/physics more.

+++ game_engine/src/game_engine/Component.java
@@ -7,7 +7,7 @@ public abstract class Component {
 	protected GameObject owner;
 	protected Constants.ComponentType compType;
 	
-	public void tick(double delta) {
+	public void tick(float delta) {
 		
 	}
